Implement LinkedList#remove_at
Removes a node at the desired index by changing previous node next pointer to the current target next pointer. If the index is 0 and it is the only node it makes head nil. If index is 0 and there are more nodes it changes the head variable to its own next pointer